+++ pom.xml
@@ -63,6 +63,7 @@
         <module>rill-flow-interfaces</module>
         <module>rill-flow-plugins</module>
         <module>rill-flow-trigger</module>
+        <module>rill-flow-task-template</module>
     </modules>
 
     <properties>
@@ -104,6 +105,11 @@
             <dependency>
                 <groupId>com.weibo</groupId>
                 <artifactId>rill-flow-service</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.weibo</groupId>
+                <artifactId>rill-flow-task-template</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
@@ -233,6 +239,24 @@
                 <version>2.9.11</version>
             </dependency>
             <!--spring end -->
+
+            <!--mybatis start-->
+            <dependency>
+                <groupId>mysql</groupId>
+                <artifactId>mysql-connector-java</artifactId>
+                <version>6.0.6</version>
+            </dependency>
+            <dependency>
+                <groupId>org.mybatis</groupId>
+                <artifactId>mybatis-spring</artifactId>
+                <version>2.0.6</version>
+            </dependency>
+            <dependency>
+                <groupId>org.mybatis</groupId>
+                <artifactId>mybatis</artifactId>
+                <version>3.5.7</version>
+            </dependency>
+            <!--mybatis end-->
 
             <!--swagger start-->
             <dependency>
