+++ rill-flow-task-template/src/main/java/com/weibo/rill/flow/task/template/model/TaskTemplateParams.java
@@ -1,0 +1,40 @@
+/*
+ *  Copyright 2021-2023 Weibo, Inc.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package com.weibo.rill.flow.task.template.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * 作为查询 TaskTemplateDO 时的参数
+ */
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
+public class TaskTemplateParams {
+    private Long id;
+    private String name;
+    private String category;
+    private Integer type;
+    private String nodeType;
+    private Integer enable;
+    private int offset;
+    private int limit;
+}
