+++ rill-flow-web/src/main/resources/spring/web.xml
@@ -21,6 +21,6 @@
 
     <import resource="classpath:spring/web-mvc.xml"/>
     <import resource="classpath:spring/service-httpclient.xml"/>
-    <import resource="service-mysql.xml"/>
+    <import resource="classpath:spring/service-mysql.xml"/>
 
 </beans>
